backend/requirements.txt
redis
python-etcd
